Added 'safe' versions of malloc(), calloc(), and realloc() to utils, and replaced all regular ones with the 'safe' ones.
The 'safe' version of these functions just handle allocation failures by printing an error and exiting the program if the return value is null.

+++ AdventOfCode_C/AoC2025_Day1.c
@@ -79,7 +79,7 @@ static void IncrementPasswordNewMethod(char *instruction, int *password, int *di
     *password += rotationFullCycles;
 
     *dial += rotation;
-    *dial = Modulo(*dial, CLICK_COUNT);
+    *dial = (int)Modulo(*dial, CLICK_COUNT);
 }
 
 /// @brief Prcoesses all rotation instructions in the given file and returns the password from it.
